--- src/main/kotlin/omok/model/Stone.kt
@@ -1,9 +1,0 @@
-package omok.model
-
-interface Stone {
-//    val color: StoneColor
-
-    fun position(): Position
-
-    fun color(): StoneColor
-}
