--- src/main/kotlin/omok/model/StoneState.kt
@@ -1,7 +1,0 @@
-package omok.model
-
-enum class StoneState {
-    Black,
-    White,
-    NONE,
-}
