+++ src/main/kotlin/omok/model/board/Position.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.board
 
 data class Position(
     val x: X,
