+++ src/main/kotlin/omok/model/board/X.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.board
 
 @JvmInline
 value class X(
