+++ src/main/kotlin/omok/model/stone/Stone.kt
@@ -1,0 +1,9 @@
+package omok.model.stone
+
+import omok.model.board.Position
+
+interface Stone {
+    fun position(): Position
+
+    fun color(): StoneColor
+}
