+++ src/main/kotlin/omok/model/stone/StoneColor.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.stone
 
 enum class StoneColor {
     BLACK,
