+++ src/main/kotlin/omok/model/stone/StoneState.kt
@@ -1,0 +1,7 @@
+package omok.model.stone
+
+enum class StoneState {
+    BLACK,
+    NONE,
+    WHITE,
+}
